// src/app/modules/pages/models/Ficticio.ts
import { Time } from "@angular/common";

export class Ficticio {
  id?: number;
  idcarrera: number;
  nombreempresa: string;
  nombrearea: string;
  nombre_prficticio: string;
  estudianes_requeridos: number;
  horas_cumplir: number;
  fecha_inicio: Date;
  actividades?: string;
  requerimientos?: string;
  estadoficticio: number;
  created_at?: Date;
  updated_at?: Date;
}